Add GH action and improve tests
