+++ .github/workflows/test-extension.yml
@@ -1,0 +1,70 @@
+name: Tests extension
+on:
+  - push
+
+jobs:
+  build:
+    runs-on: ubuntu-latest
+
+    container:
+      image: ckan/ckan-dev:2.11.3
+
+    services:
+      solr:
+        image: ckan/ckan-solr:2.10-spatial
+
+      postgres:
+        image: ckan/ckan-postgres-dev:2.11
+        env:
+          POSTGRES_USER: postgres
+          POSTGRES_PASSWORD: postgres
+          POSTGRES_DB: postgres
+        options: --health-cmd pg_isready --health-interval 10s --health-timeout 5s --health-retries 5
+
+      redis:
+        image: redis
+        options: >-
+          --health-cmd "redis-cli ping"
+          --health-interval 10s
+          --health-timeout 5s
+          --health-retries 5
+
+    env:
+      CKAN_SQLALCHEMY_URL: postgresql://ckan_default:pass@postgres/ckan_test
+      CKAN_DATASTORE_WRITE_URL: postgresql://datastore_write:pass@postgres/datastore_test
+      CKAN_DATASTORE_READ_URL: postgresql://datastore_read:pass@postgres/datastore_test
+      CKAN_SOLR_URL: http://solr:8983/solr/ckan
+      CKAN_REDIS_URL: redis://redis:6379/1
+
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v4
+
+      # - name: Install OS dependencies
+      #   run: |
+      #     apt install postgresql-client
+
+      - name: Install Python dependencies
+        run: |
+          pip install -r dev-requirements.txt
+          pip install -r requirements.txt
+          pip install -e .
+
+      - name: flake8
+        run: |
+          flake8 ckanext --count --show-source --max-complexity=14 --max-line-length=130 --statistics
+
+      - name: Setup extension
+        shell: bash
+        run: |
+          # use = config:/srv/app/src/ckan/test-core.ini
+          # Replace default path to CKAN core config file with the one on the container
+          sed -i -e 's/use = config:.*/use = config:\/srv\/app\/src\/ckan\/test-core.ini/' test.ini
+          ckan -c test.ini db init
+          # echo "Applying migrations"
+          # ckan -c test.ini db upgrade -p EXT_NAME
+
+      - name: Run tests
+        shell: bash
+        run: |
+          pytest --ckan-ini=test.ini -vv --disable-warnings ckanext/dbquery
