+++ ckanext/dbquery/tests/test_plugin.py
@@ -1,57 +1,46 @@
-"""
-Tests for plugin.py.
 
-Tests are written using the pytest library (https://docs.pytest.org), and you
-should read the testing guidelines in the CKAN docs:
-https://docs.ckan.org/en/2.9/contributing/testing.html
-
-To write tests for your extension you should install the pytest-ckan package:
-
-    pip install pytest-ckan
-
-This will allow you to use CKAN specific fixtures on your tests.
-
-For instance, if your test involves database access you can use `clean_db` to
-reset the database:
-
-    import pytest
-
-    from ckan.tests import factories
-
-    @pytest.mark.usefixtures("clean_db")
-    def test_some_action():
-
-        dataset = factories.Dataset()
-
-        # ...
-
-For functional tests that involve requests to the application, you can use the
-`app` fixture:
-
-    from ckan.plugins import toolkit
-
-    def test_some_endpoint(app):
-
-        url = toolkit.url_for('myblueprint.some_endpoint')
-
-        response = app.get(url)
-
-        assert response.status_code == 200
-
-
-To temporary patch the CKAN configuration for the duration of a test you can use:
-
-    import pytest
-
-    @pytest.mark.ckan_config("ckanext.myext.some_key", "some_value")
-    def test_some_action():
-        pass
-"""
 import pytest
 from ckan import plugins
+
+from ckanext.dbquery.plugin import DbqueryPlugin
 
 
 @pytest.mark.ckan_config("ckan.plugins", "dbquery")
 @pytest.mark.usefixtures("with_plugins")
 def test_plugin():
     assert plugins.plugin_loaded("dbquery")
+
+
+class TestDbqueryPlugin:
+
+    def test_plugin_initialization(self):
+        """Test that the plugin initializes correctly."""
+        plugin = DbqueryPlugin()
+        assert plugin is not None
+
+    def test_get_blueprint(self):
+        """Test that get_blueprint returns a blueprint."""
+        plugin = DbqueryPlugin()
+        blueprints = plugin.get_blueprint()
+
+        assert blueprints is not None
+        assert isinstance(blueprints, list)
+        assert any(b.name == 'dbquery' and b.url_prefix == '/ckan-admin/db-query' for b in blueprints)
+
+    def test_get_actions(self):
+        """Test that get_actions returns the expected actions."""
+        plugin = DbqueryPlugin()
+        actions = plugin.get_actions()
+
+        assert 'query_database' in actions
+        assert 'dbquery_executed_list' in actions
+        assert callable(actions['query_database'])
+        assert callable(actions['dbquery_executed_list'])
+
+    def test_get_auth_functions(self):
+        """Test that get_auth_functions returns the expected auth functions."""
+        plugin = DbqueryPlugin()
+        auth_functions = plugin.get_auth_functions()
+
+        assert 'query_database' in auth_functions
+        assert callable(auth_functions['query_database'])
